Move enhanceAssert() into ./lib/enhance-assert.js
