+++ lib/assert.js
@@ -1,8 +1,5 @@
 'use strict';
 var assert = require('core-assert');
-var empower = require('empower');
-var powerAssertFormatter = require('power-assert-formatter');
-var powerAssertRenderers = require('power-assert-renderers');
 
 var x = module.exports;
 
@@ -22,33 +19,6 @@ function test(ok, opts) {
 	if (!ok) {
 		throw new assert.AssertionError(opts);
 	}
-}
-
-function enhanceAssert(assert) {
-	empower(assert,
-		powerAssertFormatter({
-			renderers: [
-				powerAssertRenderers.AssertionRenderer,
-				powerAssertRenderers.SuccinctRenderer
-			]
-		}),
-		{
-			destructive: true,
-			modifyMessageOnRethrow: true,
-			saveContextOnRethrow: false,
-			patterns: [
-				't.ok(value, [message])',
-				't.notOk(value, [message])',
-				't.true(value, [message])',
-				't.false(value, [message])',
-				't.is(value, expected, [message])',
-				't.not(value, expected, [message])',
-				't.same(value, expected, [message])',
-				't.notSame(value, expected, [message])',
-				't.regexTest(regex, contents, [message])'
-			]
-		}
-	);
 }
 
 x.pass = function (msg) {
@@ -152,4 +122,4 @@ x.ifError = x.error = function (err, msg) {
 	test(!err, create(err, 'Error', '!==', msg, x.ifError));
 };
 
-enhanceAssert(x);
+require('./enhance-assert')(x);
