+++ lib/babel.js
@@ -11,17 +11,7 @@ var options = {
 	optional: hasGenerators ? ['asyncToGenerator'] : [],
 	plugins: [
 		createEspowerPlugin(require('babel-core'), {
-			patterns: [
-				't.ok(value, [message])',
-				't.notOk(value, [message])',
-				't.true(value, [message])',
-				't.false(value, [message])',
-				't.is(value, expected, [message])',
-				't.not(value, expected, [message])',
-				't.same(value, expected, [message])',
-				't.notSame(value, expected, [message])',
-				't.regexTest(regex, contents, [message])'
-			]
+			patterns: require('./enhance-assert').PATTERNS
 		})
 	]
 };
